social_network/requirements.txt
django==5.0.2
djangorestframework==3.14.0
pillow==12.0.0
psycopg2-binary==2.9.11
geopy==2.4.1
python-dotenv==1.1.1
